REFACTORED THE CODE TO NOT USE ARRAY LIST
