+++ Employee.java
@@ -1,0 +1,9 @@
+public interface Employee {
+
+        String work ();
+        double calculateSalary();
+        String getEmployeeInfo();
+    
+    }
+
+
