+++ EmployeeManagementSystem.java
@@ -1,20 +1,19 @@
-
-import  java.util.ArrayList;
-import java.util.List;
-
 public class EmployeeManagementSystem {
 
     public static void main(String[] args) {
-        Employees.FullTimeEmployee fullTimeEmployee = new Employees().new FullTimeEmployee("Alice", "FT123", 5000);
-        Employees.PartTimeEmployees partTimeEmployee = new Employees().new PartTimeEmployees("Bob", "PT456", 20, 80);
-        
+        FullTimeEmployee fullTimeEmployee = new FullTimeEmployee("Alice", "FT123", 5000);
+        PartTimeEmployees partTimeEmployee = new PartTimeEmployees("Bob", "PT456", 20, 80);
 
-        List<Employees.Employee> employees = new ArrayList<>();
-        employees.add(fullTimeEmployee);
-        employees.add(partTimeEmployee);
-        
-       Payroll payroll = new Payroll();
-       payroll.processPayroll(employees);
+        System.out.println(fullTimeEmployee.work());
+        System.out.println("Employee Info: \n" + fullTimeEmployee.getEmployeeInfo());
+        System.out.println("Salary: " + fullTimeEmployee.calculateSalary());
+        System.out.println("-------------");
+
+
+        System.out.println(partTimeEmployee.work());
+        System.out.println("Employee Info: \n" + partTimeEmployee.getEmployeeInfo());
+        System.out.println("Salary: " + partTimeEmployee.calculateSalary());
+        System.out.println("-------------");
     }
 }
