+++ Payroll.Java
@@ -1,9 +1,8 @@
 import java.util.List;
 
 public class Payroll {
-
-    public void processPayroll(List<Employees.Employee> employees) {
-        for (Employees.Employee employee : employees) {
+    public Payroll(List<Employee> employees) {
+        for (Employee employee : Employee) {
             System.out.println(employee.work());
             System.out.println("Employee Info: \n" + employee.getEmployeeInfo());
             System.out.println("Salary: " + employee.calculateSalary());
